requirements.txt
# Dependencies are set at the lowest currently tested version
# If you need an older version of a dependency please open an Issue on Github
# Basics
setuptools>=38.5
six>=1.11.0

# Version comparison
packaging>=17.1

# Templates management implementation
lxml>=4.1.1

# Security and Connectivity
requests[security]>=2.18
# urllib3 is handled by requests

# Import Export and Utils implementation
ruamel.yaml==0.15.94
deepdiff>=3.3.0,<4.0  # pyup: ignore

# Demo deployment automation
docker>=2.5.1

# Web UI
django>=2.2.6
djangorestframework>=3.10.3
django-background-tasks>=1.2.0
google-cloud-container>=0.3.0
